consolidate SepParseMask
